added testing libraries and added tests

+++ backend/routes/routes.js
@@ -2,6 +2,10 @@ const express = require("express");
 const router = express.Router();
 const db = require("../db/dbHelper");
 
+router.get("/", (req, res) => {
+  res.status(201).json("up");
+});
+
 router.get("/notes", async (req, res) => {
   try {
     const allNotes = await db.getNotes();
@@ -14,13 +18,16 @@ router.get("/notes", async (req, res) => {
 router.post("/notes/add", async (req, res) => {
   try {
     const { title, body } = req.body;
+
     if (title && body) {
       const savedNote = await db.addNote(req.body);
 
       res.status(201).json({ savedNote });
+    } else {
+      res.status(501).json({ message: "Title and Body required" });
     }
   } catch (err) {
-    res.status(500).json({ message: err });
+    res.status(501).json({ message: err });
   }
 });
 
@@ -33,8 +40,10 @@ router.put("/notes/:id", async (req, res) => {
       if (updatedNote.n === 1) {
         return res.status(201).json({ updatedNote });
       } else {
-        res.status(500).json({ message: updatedNote });
+        res.status(500).json({ message: "Title and Body required" });
       }
+    } else {
+      res.status(500).json({ message: err });
     }
   } catch (err) {
     res.status(500).json({ message: err });
@@ -44,12 +53,16 @@ router.put("/notes/:id", async (req, res) => {
 router.delete("/notes/:id", async (req, res) => {
   try {
     const { id } = req.params;
-    const deletedCount = await db.deleteNote(id);
+    if (id) {
+      const deletedCount = await db.deleteNote(id);
 
-    if (deletedCount.n === 1) {
-      res.status(201).json(deletedCount.n);
+      if (deletedCount.n === 1) {
+        res.status(201).json(deletedCount.n);
+      } else {
+        res.status(501).json({ message: deletedCount });
+      }
     } else {
-      res.status(501).json({ message: deletedCount });
+      res.status(501).json({ message: "Note ID requred" });
     }
   } catch (err) {
     res.status(501).json({ message: err });
